+++ libsolidity/formal/SMTEncoder.cpp
@@ -152,10 +152,12 @@ void SMTEncoder::visitFunctionOrModifier()
 
 	if (m_modifierDepthStack.back() == static_cast<int>(function.modifiers().size()))
 	{
-		pushPathCondition(currentPathConditions());
 		if (function.isImplemented())
+		{
+			pushInlineFrame(function);
 			function.body().accept(*this);
-		popPathCondition();
+			popInlineFrame(function);
+		}
 	}
 	else
 	{
@@ -192,7 +194,7 @@ void SMTEncoder::inlineModifierInvocation(ModifierInvocation const* _invocation,
 	initializeFunctionCallParameters(*_definition, args);
 
 	pushCallStack({_definition, _invocation});
-	pushPathCondition(currentPathConditions());
+	pushInlineFrame(*_definition);
 	if (auto modifier = dynamic_cast<ModifierDefinition const*>(_definition))
 	{
 		if (modifier->isImplemented())
@@ -205,7 +207,7 @@ void SMTEncoder::inlineModifierInvocation(ModifierInvocation const* _invocation,
 			function->accept(*this);
 		// Functions are popped from the callstack in endVisit(FunctionDefinition)
 	}
-	popPathCondition();
+	popInlineFrame(*_definition);
 }
 
 void SMTEncoder::inlineConstructorHierarchy(ContractDefinition const& _contract)
@@ -287,6 +289,16 @@ bool SMTEncoder::visit(TryCatchClause const& _clause)
 		"Assertion checker does not support try/catch clauses."
 	);
 	return false;
+}
+
+void SMTEncoder::pushInlineFrame(CallableDeclaration const&)
+{
+	pushPathCondition(currentPathConditions());
+}
+
+void SMTEncoder::popInlineFrame(CallableDeclaration const&)
+{
+	popPathCondition();
 }
 
 void SMTEncoder::endVisit(VariableDeclarationStatement const& _varDecl)
@@ -942,16 +954,21 @@ void SMTEncoder::visitTypeConversion(FunctionCall const& _funCall)
 	solAssert(_funCall.arguments().size() == 1, "");
 
 	auto argument = _funCall.arguments().front();
-	auto const& argType = argument->annotation().type;
+	auto const argType = argument->annotation().type;
+	auto const funCallType = _funCall.annotation().type;
 
-	unsigned argSize = argument->annotation().type->storageBytes();
-	unsigned castSize = _funCall.annotation().type->storageBytes();
+	auto symbArg = expr(*argument, funCallType);
 
-	auto const& funCallType = _funCall.annotation().type;
+	if (smt::isStringLiteral(*argType) && smt::isFixedBytes(*funCallType))
+	{
+		defineExpr(_funCall, symbArg);
+		return;
+	}
 
 	// TODO Simplify this whole thing for 0.8.0 where weird casts are disallowed.
 
-	auto symbArg = expr(*argument, funCallType);
+	unsigned argSize = argType->storageBytes();
+	unsigned castSize = funCallType->storageBytes();
 	bool castIsSigned = smt::isNumber(*funCallType) && smt::isSigned(funCallType);
 	bool argIsSigned = smt::isNumber(*argType) && smt::isSigned(argType);
 	optional<smtutil::Expression> symbMin;
@@ -1095,7 +1112,7 @@ void SMTEncoder::endVisit(Literal const& _literal)
 
 		addArrayLiteralAssertions(
 			*symbArray,
-			applyMap(_literal.value(), [&](auto const& c) { return smtutil::Expression{size_t(c)}; })
+			applyMap(_literal.value(), [](unsigned char c) { return smtutil::Expression{size_t(c)}; })
 		);
 	}
 	else
@@ -1952,12 +1969,6 @@ smtutil::Expression SMTEncoder::compoundAssignment(Assignment const& _assignment
 	solAssert(compoundToArithmetic.count(op) || compoundToBitwise.count(op), "");
 
 	auto decl = identifierToVariable(_assignment.leftHandSide());
-
-	TypePointer commonType = Type::commonType(
-		_assignment.leftHandSide().annotation().type,
-		_assignment.rightHandSide().annotation().type
-	);
-	solAssert(commonType == _assignment.annotation().type, "");
 
 	if (compoundToBitwise.count(op))
 		return bitwiseOperation(